fixing lint errors

+++ include/can/core/messages.hpp
@@ -920,7 +920,8 @@ struct PeripheralInfoResponse
 
     template <bit_utils::ByteIterator Output, typename Limit>
     auto serialize(Output body, Limit limit) const -> uint8_t {
-        auto iter = bit_utils::int_to_bytes(sensor, body, limit);
+        auto iter =
+            bit_utils::int_to_bytes(static_cast<uint8_t>(sensor), body, limit);
         iter = bit_utils::int_to_bytes(status, body, limit);
         return iter - body;
     }
